Make librados pool_create respect default_crush_ruleset
Make pool_create in librados use the osd_pool_default_crush_replicated_ruleset

Fixes: #11640

+++ src/messages/MPoolOp.h
@@ -30,7 +30,7 @@ public:
   __u32 op;
   uint64_t auid;
   snapid_t snapid;
-  __s16 crush_rule;
+  int crush_rule;
 
   MPoolOp()
     : PaxosServiceMessage(CEPH_MSG_POOLOP, 0, HEAD_VERSION, COMPAT_VERSION) { }
